LayoutTests/imported/w3c/web-platform-tests/largest-contentful-paint/web-font-styled-text-resize-swap-after-interaction-expected.txt


LCP: Web Font Styled Text Resize

FAIL LCP should be not updated if the web font styled text resize occurs after an interaction happens assert_equals: Web font should be downloaded. expected 1 but got 0
